Minor updates, warning fixes etc.
README.md with information on licensing and building.

fix minor compiler warning on mismatching types

Makefile: Switch to -O3 optimization level for additional speed.
Makefile: Add profile target

l3math.c: compact round_to_int without branches

mp3enc.cpp: Minor code cleanup for get_bitrate-functions

rename "rpsl.txt" to "RPSL.txt" match license text

tomp3.cpp: Enable Xing TOC by default
tomp3.cpp: Minor help text improvements

update source code modification dates

+++ hmp3/src/setup.c
@@ -1,5 +1,5 @@
 /* ***** BEGIN LICENSE BLOCK *****  
- * Source last modified: $Id: setup.c,v 1.2 2005/08/09 20:43:42 karll Exp $ 
+ * Source last modified: 2022/11/23, Maik Merten
  *   
  * Portions Copyright (c) 1995-2005 RealNetworks, Inc. All Rights Reserved.  
  *       
@@ -322,7 +322,7 @@ out_setup ( MPEG_HEAD * h )     /* info only */
     fprintf(stderr, "\n Layer %s ", Layer_msg[h->option] );
 
     fprintf(stderr, "  %s ", mode_msg[h->mode] );
-    fprintf(stderr, "  %ldkHz ", sr_table[4 * h->id + h->sr_index] );
+    fprintf(stderr, "  %ldkHz ", (long int) sr_table[4 * h->id + h->sr_index] );
     fprintf(stderr, "  %dkbps ", bitrate );
     if ( h->mode == 1 )
         fprintf(stderr, "  %d stereo bands ", 4 + 4 * h->mode_ext );
@@ -343,7 +343,7 @@ mpeg_info_string ( MPEG_HEAD * h, char *s, E_CONTROL * ec )
     if ( ( h->mode == 1 ) && ( ec->nsbstereo < 32 ) )
         s += sprintf ( s, " IS-%d ", ec->nsbstereo );
 
-    s += sprintf ( s, "  %ldHz ", sr_table[4 * h->id + h->sr_index] );
+    s += sprintf ( s, "  %ldHz ", (long int) sr_table[4 * h->id + h->sr_index] );
 
     if ( ec->vbr_flag == 0 )
     {
